docs(sandbox): split into multi-page section and reframe as an interface
Address review feedback on PR #2851:

- Split the single page into a section (Overview, Available Sandboxes,
  Building a Custom Sandbox), matching how Tools and Model Providers are
  organized.
- Reframe the sandbox as an interface the agent uses to reach an execution
  environment you provision, not a wrapper that sandboxes the agent process.
  Add a 'Two Ways to Run an Agent With a Sandbox' section using the
  harness/compute (control plane vs. execution plane) framing from OpenAI's
  sandbox agents guide.
- Rework Security to state plainly that the guarantees depend on how you
  configure the environment behind the interface.
- Update navigation.yml to a labeled Sandbox section with subpages.
- Split TypeScript snippet files per page.
