+++ site/src/content/docs/user-guide/concepts/sandbox/available-sandboxes_imports.ts
@@ -1,0 +1,22 @@
+// @ts-nocheck
+// Imports for the Available Sandboxes page snippets.
+
+// --8<-- [start:docker_imports]
+import { Agent } from '@strands-agents/sdk'
+import { DockerSandbox } from '@strands-agents/sdk/sandbox/docker'
+// --8<-- [end:docker_imports]
+
+// --8<-- [start:ssh_imports]
+import { Agent } from '@strands-agents/sdk'
+import { SshSandbox } from '@strands-agents/sdk/sandbox/ssh'
+// --8<-- [end:ssh_imports]
+
+// --8<-- [start:direct_use_imports]
+import { Agent } from '@strands-agents/sdk'
+import { DockerSandbox } from '@strands-agents/sdk/sandbox/docker'
+// --8<-- [end:direct_use_imports]
+
+// --8<-- [start:streaming_imports]
+import { Agent } from '@strands-agents/sdk'
+import { DockerSandbox } from '@strands-agents/sdk/sandbox/docker'
+// --8<-- [end:streaming_imports]
