+++ site/src/content/docs/user-guide/concepts/sandbox/index.ts
@@ -1,12 +1,6 @@
 import { Agent } from '@strands-agents/sdk'
 import { DockerSandbox } from '@strands-agents/sdk/sandbox/docker'
-import { SshSandbox } from '@strands-agents/sdk/sandbox/ssh'
-import { PosixShellSandbox } from '@strands-agents/sdk/sandbox'
-import type { ExecuteOptions, ExecutionResult, StreamChunk } from '@strands-agents/sdk/sandbox'
-import type { Tool } from '@strands-agents/sdk'
 import { makeBash } from '@strands-agents/sdk/vended-tools/bash'
-import { makeFileEditor } from '@strands-agents/sdk/vended-tools/file-editor'
-import { spawn } from 'node:child_process'
 
 async function configuring() {
   // --8<-- [start:configuring]
@@ -51,107 +45,6 @@ function overrideTool() {
   // --8<-- [end:override_tool]
 }
 
-function docker() {
-  // --8<-- [start:docker]
-  // workingDir and user are optional; omit them to use the container's defaults
-  const sandbox = new DockerSandbox({
-    container: 'agent-workspace',
-    workingDir: '/workspace',
-    user: '1000:1000',
-  })
-  const agent = new Agent({ sandbox })
-  void agent
-  // --8<-- [end:docker]
-}
-
-function ssh() {
-  // --8<-- [start:ssh]
-  const sandbox = new SshSandbox({
-    host: 'ubuntu@10.0.1.5',
-    workingDir: '/home/ubuntu/workspace',
-    identityFile: '~/.ssh/agent_key',
-  })
-  const agent = new Agent({ sandbox })
-  void agent
-  // --8<-- [end:ssh]
-}
-
-async function directUse() {
-  // --8<-- [start:direct_use]
-  const agent = new Agent({
-    sandbox: new DockerSandbox({ container: 'agent-workspace', workingDir: '/workspace' }),
-  })
-
-  // Seed an input file, let the agent work, then read the result back
-  await agent.sandbox.writeText('/workspace/input.csv', 'id,value\n1,42\n')
-
-  await agent.invoke('Summarize /workspace/input.csv and write the summary to /workspace/out.txt')
-
-  const result = await agent.sandbox.execute('cat /workspace/out.txt')
-  console.log(result.exitCode, result.stdout)
-  // --8<-- [end:direct_use]
-}
-
-async function streaming() {
-  // --8<-- [start:streaming]
-  const agent = new Agent({
-    sandbox: new DockerSandbox({ container: 'agent-workspace', workingDir: '/workspace' }),
-  })
-
-  for await (const chunk of agent.sandbox.executeStreaming('npm run build')) {
-    if (chunk.type === 'streamChunk') {
-      process.stdout.write(chunk.data)
-    } else {
-      console.log(`\nexit code: ${chunk.exitCode}`)
-    }
-  }
-  // --8<-- [end:streaming]
-}
-
-// --8<-- [start:custom_posix]
-class FirecrackerSandbox extends PosixShellSandbox {
-  constructor(private readonly vmId: string) {
-    super()
-  }
-
-  async *executeStreaming(
-    command: string,
-    options?: ExecuteOptions
-  ): AsyncGenerator<StreamChunk | ExecutionResult, void, undefined> {
-    void options
-    const proc = spawn('fc-exec', [this.vmId, 'sh', '-c', command])
-
-    let stdout = ''
-    let stderr = ''
-    for await (const data of proc.stdout) {
-      const text = data.toString()
-      stdout += text
-      yield { type: 'streamChunk', data: text, streamType: 'stdout' }
-    }
-    for await (const data of proc.stderr) {
-      const text = data.toString()
-      stderr += text
-      yield { type: 'streamChunk', data: text, streamType: 'stderr' }
-    }
-    const exitCode: number = await new Promise((resolve) => proc.on('close', (code) => resolve(code ?? 0)))
-    yield { type: 'executionResult', exitCode, stdout, stderr, outputFiles: [] }
-  }
-}
-// --8<-- [end:custom_posix]
-
-class FirecrackerSandboxWithTools extends FirecrackerSandbox {
-  // --8<-- [start:custom_tools]
-  override getTools(): Tool[] {
-    return [makeFileEditor(this, { name: 'sandbox_file_editor' }), makeBash(this, { name: 'sandbox_bash' })]
-  }
-  // --8<-- [end:custom_tools]
-}
-
 void configuring
 void hostDefault
 void overrideTool
-void docker
-void ssh
-void directUse
-void streaming
-void FirecrackerSandboxWithTools
